+++ app/services/parkService.js
@@ -2,11 +2,10 @@
 
 angular.module('national-parks')
   .factory('parkFactory', ($http) => {
-    const numPhotos = 10;
     let parkFactory = {};
 
     parkFactory.fetchPhotos = (parkName) => {
-      let data = $http.get('/park/flickr?parkName=' + parkName + '&numPhotos=' + numPhotos)
+      let data = $http.get('/park/flickr?parkName=' + parkName)
         .then((response) => {
           return response.data.body.photos.photo;
         }, (error) => {
