+++ app/styles/style.css
@@ -14,6 +14,10 @@ footer {
   margin-bottom: 20px;
 }
 
+footer > p {
+  float: right !important;
+}
+
 .park {
   display: flex;
   flex-direction: column;
@@ -57,9 +61,25 @@ footer {
   flex: 1;
 }
 
-#map {
+.park-hero {
   width: 100%;
-  height: 60vh;
+  height: 360px !important;
+  object-fit: cover;
+}
+
+.panel {
+  margin-bottom: 0;
+}
+
+.park-col {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+#parkMap {
+  width: 100%;
+  height: 360px;
 }
 
 @media (max-width: 768px) {
@@ -112,6 +132,10 @@ footer {
 
   .foursquare-and-distance > * {
     font-size: 1em;
+  }
+
+  .park-row {
+    display: flex;
   }
 
 }
