+++ app/views/footer.html
@@ -1,4 +1,4 @@
 <footer class="footer">
   <hr  />
-  Made by <a href="http://www.alanseto.com/" target="_blank">Alan Seto</a>
+  <p>Made by <a href="http://www.alanseto.com/" target="_blank">Alan Seto</a></p>
 </footer>
