+++ app/views/park.html
@@ -1,12 +1,63 @@
-<div class="row">
-  <div class="col-xs-12">
-    <a ui-sref="app" class="btn btn-info btn-map" role="button" style="margin-bottom:20px;">Back</a>
+<div class="row park-row" style="border:1px solid red;">
+  <div class="park-col col-sm-12 col-md-8" style="border:1px solid green;">
+    <div id="parkCarousel" class="carousel slide" data-ride="carousel">
+      <div class="carousel-inner" role="listbox">
+        <div class="item park-hero-div" ng-class="{'active': $first}" ng-repeat="photo in park.flickrPhotos">
+          <img class="park-hero" ng-src="{{photo.src}}" alt="{{photo.title}}">
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="park-col col-sm-12 col-md-4" style="border:1px solid green;">
+    <h2 class="text-center">{{park.fullName}}</h2>
+    <div class="panel panel-default">
+      <div class="panel-body">
+        <p>{{park.description}}</p>
+      </div>
+    </div>
   </div>
 </div>
-<div id="parkCarousel" class="carousel slide" data-ride="carousel">
-  <div class="carousel-inner" role="listbox">
-    <div class="item" ng-class="{'active': $first}" ng-repeat="photo in park.photos">
-      <img class="{{$index}}-slide" ng-src="{{photo.src}}" alt="{{photo.title}}" style="height:750px;width:100%;">
+<div class="row park-row" style="border:1px solid red;">
+  <div class="park-col col-sm-12 col-md-4" style="border:1px solid green;">
+    <div class="panel panel-default">
+      <div class="panel-heading">
+        <h3 class="panel-title">Foursquare</h3>
+      </div>
+      <div class="panel-body">
+        <h3>Rating: {{park.foursquare.venue.rating}}</h3>
+        <h3># of Ratings: {{park.foursquare.venue.ratingSignals}}</h3>
+        <h3>Checkins: {{park.foursquare.venue.stats.checkinsCount}}</h3>
+      </div>
     </div>
+  </div>
+  <div class="park-col col-sm-12 col-md-8" style="border:1px solid green;">
+    <div class="panel panel-default">
+      <div class="panel-heading">
+        <h3 class="panel-title">Location</h3>
+      </div>
+      <div id="parkMap"></div>
+      <div class="panel-footer">
+        {{park.directionsInfo}}
+      </div>
+    </div>
+  </div>
+</div>
+<div class="row park-row" style="border:1px solid red;">
+  <div class="park-colcol-sm-12 col-md-8" style="border:1px solid green;">
+    <div class="panel panel-default">
+      <div class="panel-heading">
+        <h3 class="panel-title">Weather</h3>
+      </div>
+      <div class="panel-body">
+
+      </div>
+      <div class="panel-footer">
+        {{park.weatherInfo}}
+      </div>
+    </div>
+  </div>
+  <div class="park-col col-sm-12 col-md-4" style="border:1px solid green;">
+    <p><a href="{{park.directionsUrl}}">Directions URL</a></p>
+    <p><a href="{{park.url}}">Park URL</a></p>
   </div>
 </div>
