+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "national-parks",
-  "version": "0.11.7",
+  "version": "0.12.0",
   "description": "United States National Parks",
   "main": "index.html",
   "license": "MIT",
